@@ -1,7 +1,7 @@
 package com.kuehne_nagel.city_list.application.transport.request;
 
-import javax.validation.Valid;
-import javax.validation.constraints.NotNull;
+import jakarta.validation.Valid;
+import jakarta.validation.constraints.NotNull;
 
 import lombok.Data;
 
