@@ -1,6 +1,6 @@
 package com.kuehne_nagel.city_list.application.transport.request;
 
-import javax.validation.constraints.NotNull;
+import jakarta.validation.constraints.NotNull;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import lombok.Data;
