@@ -1,32 +1,21 @@
 package com.kuehne_nagel.city_list.application.config;
 
-import java.util.Collections;
-
+import org.springdoc.core.models.GroupedOpenApi;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
-import org.springframework.web.util.UriComponentsBuilder;
-import springfox.documentation.builders.PathSelectors;
-import springfox.documentation.builders.RequestHandlerSelectors;
-import springfox.documentation.service.ApiInfo;
-import springfox.documentation.service.Contact;
-import springfox.documentation.spi.DocumentationType;
-import springfox.documentation.spring.web.paths.AbstractPathProvider;
-import springfox.documentation.spring.web.paths.Paths;
-import springfox.documentation.spring.web.plugins.Docket;
-import springfox.documentation.swagger2.annotations.EnableSwagger2;
 
 @Configuration
-@EnableSwagger2
+//@EnableSwagger2
 public class SwaggerConfig {
 
-    @Value( "${base-url.context}" )
+    @Value("${base-url.context}")
     private String context;
 
-    @Value( "${app.host}" )
+    @Value("${app.host}")
     private String host;
 
-    @Bean
+    /*@Bean
     public Docket api() {
         return new Docket(DocumentationType.SWAGGER_2)
                 .host(host)
@@ -34,10 +23,36 @@ public class SwaggerConfig {
                 .paths(PathSelectors.any())
                 .build()
                 .apiInfo(apiInfo())
+                .securityContexts(Lists.newArrayList(securityContext()))
                 .pathProvider(new BasePathAwareRelativePathProvider(context));
+        return new Docket(DocumentationType.SWAGGER_2)
+                .select()
+                .apis(RequestHandlerSelectors.basePackage("com.kuehne_nagel.city_list.application.controller"))
+                .paths(PathSelectors.regex("/com/v1.0/citymgt.*"))
+                .build()
+                .groupName("city-list")
+                .apiInfo(apiInfo());
+    }*/
+    @Bean
+    public GroupedOpenApi publicApi() {
+        return GroupedOpenApi.builder()
+                .group("city-list")
+                .pathsToMatch(String.format("%s/**", context))
+                .build();
     }
+   /* private SecurityContext securityContext() {
+        return SecurityContext
+                .builder()
+                .securityReferences(defaultAuth())
+                .build();
+    }*/
 
-    private ApiInfo apiInfo() {
+   /* private List<SecurityReference> defaultAuth() {
+        AuthorizationScope[] authorizationScopes = {new AuthorizationScope("global", "accessEverything")};
+        return Lists.newArrayList(new SecurityReference("JWT", authorizationScopes));
+    }*/
+
+  /*  private ApiInfo apiInfo() {
         String title = "Test";
         return new ApiInfo(
                 title,
@@ -46,9 +61,9 @@ public class SwaggerConfig {
                 "Test",
                 new Contact("Admin", "http://www.google.com", "test@google.com"),
                 "License of API", "API license URL", Collections.emptyList());
-    }
+    }*/
 
-    class BasePathAwareRelativePathProvider extends AbstractPathProvider {
+  /*  class BasePathAwareRelativePathProvider extends AbstractPathProvider {
 
         private String basePath;
 
@@ -72,6 +87,6 @@ public class SwaggerConfig {
                     uriComponentsBuilder.path(operationPath.replaceFirst(basePath, "")).build().toString());
         }
 
-    }
+    }*/
 
 }
