@@ -1,17 +1,15 @@
 package com.kuehne_nagel.city_list.application.transport.request;
 
-import jakarta.validation.Valid;
-import jakarta.validation.constraints.NotNull;
-
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import lombok.Data;
 
+import jakarta.validation.Valid;
+import jakarta.validation.constraints.NotNull;
+
 @Data
 @JsonIgnoreProperties( ignoreUnknown = true )
-public class UpdateRequest <T> extends IdRequest {
-
-    @NotNull(message = "updateDto is null")
+public class IdRequest extends BaseRequest{
+    @NotNull(message = "id is null")
     @Valid
-    private T updateDto;
-
+    private Long id;
 }
